requirements.txt
et-xmlfile==2.0.0
lxml==6.0.0
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.0
pillow==11.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
qrcode==8.2
six==1.17.0
typing-extensions==4.14.0
tzdata==2025.2
